Moved the comment author info component to a separate file
